Update README to clarify the crawler start-up step
The first time I ran this, the documentation just said that this step would take "a while". Being ignorant of how the crawler worked, since  I was still seeing log output for hours, I thought it was still setting up and wasted a day waiting for the start-up process to terminate :-O oops. This change would keep that from happening to anyone else.