Fix an internal error claiming that the cursor must be an object Many widgets were throwing an internal error, which looked like this:
One of the dashboard widgets ("usersWithoutOrg", running against
  the "null" root had a problem, so I've skipped over it. This is
  really an internal error, and should be reported.
  (The error is described like this, which will help in the report:
  cursor options must be an object,
  node_modules/mongodb-core/lib/error.js:31)

This is because of changes in MongoDB. As https://docs.mongodb.com/manual/reference/command/aggregate/ explains:

MongoDB 3.6 removes the use of aggregate command without the cursor option unless the command includes the explain option. Unless you include the explain option, you must specify the cursor option.

    To indicate a cursor with the default batch size, specify cursor: {}.
    To indicate a cursor with a non-default batch size, use cursor: { batchSize: <num> }.

All these widgets used aggregate, and therefore needed to be given a cursor; this means that they did not call a callback with their results but instead with a cursor to point at those results, so we also add a call to toArray() which iterates the cursor for us to get an array of results as before, so the remainder of the code does not change.