Weekly dates need gggg not YYYY as their year format This is another example of https://github.com/MeasureOSS/Measure/commit/914d9b4bde09c3fafc7f7a0aaf40cc4e1c4944aa cropping up again. On this occasion, it was causing all the data in the First PR/issue response graphs after the beginning of 2014 to show up as empty, or as the same number for all graphs. This was because of this:
```
>>> moment('2013-12-31T04:00:00').format('YYYY-ww')
2013-01
>>> moment('2013-12-31T04:00:00').format('gggg-ww')
2014-01
```

and so the list of date labels for weeks was not strictly increasing (it went: 2013-51, 2013-52, 2013-01, 2014-02...)
This confused `fillGaps`, which loses track of the correct label at the point of discontinuity and then assumes that there's no data for that week, or any weeks after it, and so fills in the last value it saw.
Changing the formatter to use `gggg` means that the list of week dates is now increasing again without discontinuities, `fillGaps` is no longer sad, and we therefore don't throw away all the data after we hit one of these points.